[move] Unify permitted characters check
Both the Move source language and IR compilers check that the source
text contains only characters from a specific character set. Unify the
separate functions used to perform these checks and place them in the
move-command-line-common package, so that any Move command-line tool can
do the same.

Closes: #8860
