+++ language/move-command-line-common/src/lib.rs
@@ -3,6 +3,7 @@
 
 #![forbid(unsafe_code)]
 
+pub mod character_sets;
 pub mod env;
 pub mod files;
 pub mod testing;
